Add C++ headers to specs file
This is a local change (currently with a hardcoded GCC version(!)) to
allow us to switch between C++ headers in the same way that we currently
do C headers.

A future verison of this patch will clean up the hardcoded version,
either by changing where GCC expects the headers, or by making this a
build time constant.