libc: Provide definitions for newlib reent specific types
In toolchains which ship both newlib and picolibc, provide definitions
for newlib's reent.h in case this is included by a user. This is needed
since reent.h will come from newlib, but sys/config.h will be provided
by picolibc, and the former depends on the latter providing these
definitions.